+++ src/main/java/org/example/service/impl/UserServiceImpl.java
@@ -1,10 +1,9 @@
 package org.example.service.impl;
 
 import lombok.extern.slf4j.Slf4j;
-import org.example.dto.UserDTO;
-import org.example.entity.ProjectPosition;
 import org.example.entity.User;
 import org.example.exception.DuplicateUserException;
+import org.example.exception.EmptyInputException;
 import org.example.modelmapper.UserMapper;
 import org.example.repository.DepartmentRepository;
 import org.example.repository.UserRepository;
@@ -13,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Slf4j
@@ -29,25 +29,16 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User save(User user, Integer departmentId) {
-        log.info(String.format("user.save {id = %d, firstName = %s, lastName = %s}",
-                user.getId(), user.getFirstName(), user.getLastName()));
-        if (departmentId == null) {
-            throw new RuntimeException("Required parameters: departmentId");
-        }
-//        for (User u : userRepository.findAll()) {
-//            if (u.getEmail().equals(user.getEmail())) {
-//                throw new DuplicateUserException(user);
-//            }
-//        }
-//        if (departmentRepository.findById(departmentId).isPresent())
-        user.setDepartment(departmentRepository.getById(departmentId));
-        return userRepository.save(user);
-    }
-
-    @Override
     public User save(User user) {
         log.info(String.format("user.save {id = %d}", user.getId()));
+        if (user.getFirstName().isEmpty() || user.getFirstName().length() == 0) {
+            throw new EmptyInputException("Input fields are empty");
+        }
+        for (User u : userRepository.findAll()) {
+            if (u.getEmail().equals(user.getEmail())) {
+                throw new DuplicateUserException("User already exist");
+            }
+        }
         return userRepository.save(user);
     }
 
@@ -65,28 +56,33 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Optional<User> findById(Integer id) {
+        if (!userRepository.findById(id).isPresent()) {
+            throw new NoSuchElementException();
+        }
         return userRepository.findById(id);
     }
 
     @Override
     public User update(Integer id, User updatedUser) {
+        if (!userRepository.findById(id).isPresent()) {
+            throw new NoSuchElementException();
+        }
         User user = userRepository.findById(id).get();
         user.setFirstName(updatedUser.getFirstName());
         user.setLastName(updatedUser.getLastName());
         user.setJobTitle(updatedUser.getJobTitle());
         user.setDepartment(updatedUser.getDepartment());
         user.setEmail(updatedUser.getEmail());
-        if (!updatedUser.getPassword().isEmpty()) {
-            user.setPassword(updatedUser.getPassword());
-        }
+        user.setPassword(updatedUser.getPassword());
         user.setPassword(updatedUser.getPassword());
         return userRepository.save(user);
     }
 
     @Override
     public void deleteById(Integer id) {
-//        if (userRepository.findById(id).isPresent()) {
+        if (!userRepository.findById(id).isPresent()) {
+            throw new NoSuchElementException();
+        }
         userRepository.deleteById(id);
-//        }
     }
 }
